added button timeout after click in embedded target

+++ framework/plabel.h
@@ -26,7 +26,7 @@ protected:
 //fontprops is used for compute widget width and height
 #define DECL_LABEL_PROPERTIES(name,x,y,text,textColor,backgroundColor,font,fontprops) \
     _DECL_WIDGET_PROPERTIES(name,x,y,\
-    (fontprops.width *(sizeof(text)  - 1))\
+    (uint16_t)(fontprops.width *(sizeof(text) -1 ))\
     ,fontprops.height\
     ,text,textColor,backgroundColor,font,0)
